+++ source/ChanSort.Api/Utils/Crc16.cs
@@ -2,8 +2,11 @@
 {
   public class Crc16
   {
+    #region static IBM/Modbus implementation
+
     // CRC-16-IBM aka Modbus, LSB first (right shift) with polynomial 0x8005
-    private static readonly ushort[] CrcTable = {
+    private static readonly ushort[] CrcTable = 
+    {
        0X0000, 0XC0C1, 0XC181, 0X0140, 0XC301, 0X03C0, 0X0280, 0XC241,
        0XC601, 0X06C0, 0X0780, 0XC741, 0X0500, 0XC5C1, 0XC481, 0X0440,
        0XCC01, 0X0CC0, 0X0D80, 0XCD41, 0X0F00, 0XCFC1, 0XCE81, 0X0E40,
@@ -35,21 +38,71 @@
        0X8801, 0X48C0, 0X4980, 0X8941, 0X4B00, 0X8BC1, 0X8A81, 0X4A40,
        0X4E00, 0X8EC1, 0X8F81, 0X4F40, 0X8D01, 0X4DC0, 0X4C80, 0X8C41,
        0X4400, 0X84C1, 0X8581, 0X4540, 0X8701, 0X47C0, 0X4680, 0X8641,
-       0X8201, 0X42C0, 0X4380, 0X8341, 0X4100, 0X81C1, 0X8081, 0X4040 };
+       0X8201, 0X42C0, 0X4380, 0X8341, 0X4100, 0X81C1, 0X8081, 0X4040
+    };
 
-    public static ushort Calc(byte[] data, int startOffset = 0, int length = -1)
+    public static ushort Modbus(byte[] data, int startOffset = 0, int length = -1)
     {
       if (length == -1)
-        length = data.Length;
+        length = data.Length - startOffset;
 
       ushort crc = 0xFFFF;
       while (--length >= 0)
       {
-        byte temp = (byte)(data[startOffset++] ^ crc);
+        var temp = (byte)(data[startOffset++] ^ crc);
         crc >>= 8;
         crc ^= CrcTable[temp];
       }
 
+      return crc;
+    }
+    #endregion
+
+    #region static CCITT implementation
+
+    private static Crc16 ccitt;
+    /// <summary>
+    /// Unreflected (most significant bit first) CRC16-CCITT with polynomial 0x1021 and initial value 0xFFFF
+    /// </summary>
+    public static Crc16 CCITT => ccitt ??= new Crc16(0x1021, 0xffff);
+    #endregion
+
+    private readonly ushort[] table = new ushort[256];
+    private readonly ushort init;
+
+    public Crc16(ushort poly, ushort init = 0)
+    {
+      InitTable(poly);
+      this.init = init;
+    }
+
+    private void InitTable(ushort poly)
+    {
+      for (int i = 0; i < table.Length; ++i)
+      {
+        var temp = (ushort)0;
+        var a = (ushort)(i << 8);
+        for (int j = 0; j < 8; ++j)
+        {
+          if (((temp ^ a) & 0x8000) != 0)
+            temp = (ushort)((temp << 1) ^ poly);
+          else
+            temp <<= 1;
+          a <<= 1;
+        }
+
+        table[i] = temp;
+      }
+    }
+
+    public ushort Calc(byte[] data, int startOffset = 0, int length = -1)
+    {
+      var crc = init;
+      if (length < 0)
+        length = data.Length - startOffset;
+      var end = startOffset + length;
+      for (int i = startOffset; i < end; i++)
+        crc = (ushort)((crc << 8) ^ table[(crc >> 8) ^ (0xff & data[i])]);
       return crc;
     }
   }